Handle immediate leave transition after enter (fix #1116)
If a CSS transition leaves immediately after enter,
the transitionend event never fires because the CSS values
didn't have time to interpolate. This can cause items to
leave hanging in the DOM or cause out-in mode to never
append the entering item. By detecting that we just entered
last tick we can end a leave transition immediately.